Ajuste nos values objects para que sejam imutáveis

+++ src/main/java/br/com/fiap/tech_challenge/domain_layer/business_entities/TesteEntities.java
@@ -1,14 +1,11 @@
 package br.com.fiap.tech_challenge.domain_layer.business_entities;
 
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-
 import com.google.gson.Gson;
 
 public class TesteEntities {
     public static void main(String[] args) {
         try {
-            Cpf cpf = new Cpf(123456789, 9);
+            Cpf cpf = new Cpf(123456789, (byte) 9);
             Cliente anaJulia = new Cliente("Ana Júlia Nascimento", cpf, "anajulia@gmail.com");
 
             Gson gson = new Gson();
